% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_lovers_complaint}
\alias{a_lovers_complaint}
\title{Text of A Lover's Complaint}
\format{
A character vector of length 383. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
a_lovers_complaint
}
\description{
A character vector containing the complete text of A Lover's Complaint by
William Shakespeare.
}
\keyword{datasets}
