% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_midsummer_nights_dream}
\alias{a_midsummer_nights_dream}
\title{Text of A Midsummer Night's Dream}
\format{
A character vector of length 3460. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
a_midsummer_nights_dream
}
\description{
A character vector containing the complete text of:
A Midsummer Night's Dream
by William Shakespeare.
}
\keyword{datasets}
