% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_works_df}
\alias{all_works_df}
\title{Contents of Complete Works of William Shakespeare (dataframe)}
\format{
A data frame with 166340 rows and 4 variables:
\describe{
  \item{name}{short (or common) name of the work}
  \item{content}{the full contents of the work. Each line is ~70 characters}
  \item{full_name}{the complete name of the work, as listed}
  \item{genre}{whether the work is poetry, history, comedy, or tragedy}
}
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
all_works_df
}
\description{
A dataframe containing the full text of all of the complete works of William
Shakespeare, as provided by Project Gutenberg.
}
\examples{
works <- bardr::all_works_df
subset(works, works$genre == "History")
}
\keyword{datasets}
