% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_works_list}
\alias{all_works_list}
\title{Contents of Complete Works of William Shakespeare (list)}
\format{
A list with 44 elements, each one containing a character vector
containing the full text of a work, given in the element name.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
all_works_list
}
\description{
A list containing the full text of all of the complete works of William
Shakespeare, as provided by Project Gutenberg.
}
\keyword{datasets}
