% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alls_well_that_ends_well}
\alias{alls_well_that_ends_well}
\title{Text of All's Well That Ends Well}
\format{
A character vector of length 4962 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
alls_well_that_ends_well
}
\description{
A character vector containing the complete text of:
All's Well That Ends Well
by William Shakespeare.
}
\keyword{datasets}
