% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antony_and_cleopatra}
\alias{antony_and_cleopatra}
\title{Text of Antony and Cleopatra}
\format{
A character vector of length 4102 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
antony_and_cleopatra
}
\description{
A character vector containing the complete text of:
Antony and Cleopatra
by William Shakespeare.
}
\keyword{datasets}
