% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as_you_like_it}
\alias{as_you_like_it}
\title{Text of As You Like It}
\format{
A character vector of length 2791 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
as_you_like_it
}
\description{
A character vector containing the complete text of:
As You Like It
by William Shakespeare.
}
\keyword{datasets}
