% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coriolanus}
\alias{coriolanus}
\title{Text of Coriolanus}
\format{
A character vector of length 4117. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
coriolanus
}
\description{
A character vector containing the complete text of:
Coriolanus
by William Shakespeare.
}
\keyword{datasets}
