% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cymbeline}
\alias{cymbeline}
\title{Text of Cymbeline}
\format{
A character vector of length 5875. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
cymbeline
}
\description{
A character vector containing the complete text of:
Cymbeline
by William Shakespeare.
}
\keyword{datasets}
