% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamlet}
\alias{hamlet}
\title{Text of Hamlet}
\format{
A character vector of length 6687. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
hamlet
}
\description{
A character vector containing the complete text of:
Hamlet
by William Shakespeare.
}
\keyword{datasets}
