% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{henry_iv_part_1}
\alias{henry_iv_part_1}
\title{Text of Henry IV, Part 1}
\format{
A character vector of length 4716. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
henry_iv_part_1
}
\description{
A character vector containing the complete text of:
Henry IV, Part 1
by William Shakespeare.
}
\keyword{datasets}
