% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{henry_v}
\alias{henry_v}
\title{Text of Henry V}
\format{
A character vector of length 4936. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
henry_v
}
\description{
A character vector containing the complete text of:
Henry V
by William Shakespeare.
}
\keyword{datasets}
