% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{henry_vi_part_3}
\alias{henry_vi_part_3}
\title{Text of Henry VI, Part 3}
\format{
A character vector of length 3346. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
henry_vi_part_3
}
\description{
A character vector containing the complete text of:
A Midsummer Night's Dream
by William Shakespeare.
}
\keyword{datasets}
