% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intro}
\alias{intro}
\title{Project Gutenberg preface to Shakespeare's Complete Works}
\format{
A character vector of length 45.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
intro
}
\description{
A character vector containing all text before the beginning of the actual
text of the work.
}
\keyword{datasets}
