% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{julius_caesar}
\alias{julius_caesar}
\title{Text of Julius Caesar}
\format{
A character vector of length 4644 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
julius_caesar
}
\description{
A character vector containing the complete text of:
Julius Caesar
by William Shakespeare.
}
\keyword{datasets}
