% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{king_john}
\alias{king_john}
\title{Text of King John}
\format{
A character vector of length 2917. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
king_john
}
\description{
A character vector containing the complete text of:
King John
by William Shakespeare.
}
\keyword{datasets}
