% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{king_lear}
\alias{king_lear}
\title{Text of King Lear}
\format{
A character vector of length 6036 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
king_lear
}
\description{
A character vector containing the complete text of:
King Lear
by William Shakespeare.
}
\keyword{datasets}
