% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loves_labours_lost}
\alias{loves_labours_lost}
\title{Text of Love's Labour's Lost}
\format{
A character vector of length 2915. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
loves_labours_lost
}
\description{
A character vector containing the complete text of:
Love's Labour's Lost
by William Shakespeare.
}
\keyword{datasets}
