% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macbeth}
\alias{macbeth}
\title{Text of Macbeth}
\format{
A character vector of length 4149 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
macbeth
}
\description{
A character vector containing the complete text of:
Macbeth
by William Shakespeare.
}
\keyword{datasets}
