% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_for_measure}
\alias{measure_for_measure}
\title{Text of Measure for Measure}
\format{
A character vector of length 3000. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
measure_for_measure
}
\description{
A character vector containing the complete text of:
Measure for Measure
by William Shakespeare.
}
\keyword{datasets}
