% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{much_ado_about_nothing}
\alias{much_ado_about_nothing}
\title{Text of Much Ado About Nothing}
\format{
A character vector of length 5712. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
much_ado_about_nothing
}
\description{
A character vector containing the complete text of:
Much Ado About Nothing
by William Shakespeare.
}
\keyword{datasets}
