% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{othello}
\alias{othello}
\title{Text of Othello, the Moor of Venice}
\format{
A character vector of length 6270. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
othello
}
\description{
A character vector containing the complete text of:
Othello
by William Shakespeare.
}
\keyword{datasets}
