% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outro}
\alias{outro}
\title{Project Gutenberg ending to Shakespeare's Complete Works}
\format{
A character vector of length 416. Contains the license.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
outro
}
\description{
A character vector containing all text after the end of the text.
}
\keyword{datasets}
