% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pericles_prince_of_tyre}
\alias{pericles_prince_of_tyre}
\title{Text of Pericles, Prince of Tyre}
\format{
A character vector of length 4139 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
pericles_prince_of_tyre
}
\description{
A character vector containing the complete text of:
Pericles, Prince of Tyre
by William Shakespeare.
}
\keyword{datasets}
