% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{richard_ii}
\alias{richard_ii}
\title{Text of Richard II}
\format{
A character vector of length 3097. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
richard_ii
}
\description{
A character vector containing the complete text of:
Richard II
by William Shakespeare.
}
\keyword{datasets}
