% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{richard_iii}
\alias{richard_iii}
\title{Text of Richard III}
\format{
A character vector of length 4408. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
richard_iii
}
\description{
A character vector containing the complete text of:
Richard III
by William Shakespeare.
}
\keyword{datasets}
