% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{romeo_and_juliet}
\alias{romeo_and_juliet}
\title{Text of Romeo and Juliet}
\format{
A character vector of length 5256 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
romeo_and_juliet
}
\description{
A character vector containing the complete text of:
Romeo and Juliet
by William Shakespeare.
}
\keyword{datasets}
