% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sonnets}
\alias{sonnets}
\title{Text of Shakespeare's Sonnets}
\format{
A character vector of length 2777 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
sonnets
}
\description{
A character vector containing the complete text of all of Shakespeare's
sonnets. They are not separated and roman numerals on their own lines are
used to mark the sonnets.
}
\details{
TO-DO: include pre-tidyed sonnets dataframe.
}
\keyword{datasets}
