% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taming_of_the_shrew}
\alias{taming_of_the_shrew}
\title{Text of The Taming of the Shrew}
\format{
A character vector of length 4874 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
taming_of_the_shrew
}
\description{
A character vector containing the complete text of:
The Taming of the Shrew
by William Shakespeare.
}
\keyword{datasets}
