% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_comedy_of_errors}
\alias{the_comedy_of_errors}
\title{Text of The Comedy of Errors}
\format{
A character vector of length 3201. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_comedy_of_errors
}
\description{
A character vector containing the complete text of:
The Comedy of Errors
by William Shakespeare.
}
\keyword{datasets}
