% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_merchant_of_venice}
\alias{the_merchant_of_venice}
\title{Text of The Merchant of Venice}
\format{
A character vector of length 4168. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_merchant_of_venice
}
\description{
A character vector containing the complete text of:
The Merchant of Venice
by William Shakespeare.
}
\keyword{datasets}
