% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_merry_wives_of_windsor}
\alias{the_merry_wives_of_windsor}
\title{Text of The Merry Wives of Windsor}
\format{
A character vector of length 2959. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_merry_wives_of_windsor
}
\description{
A character vector containing the complete text of:
The Merry Wives of Windsor
by William Shakespeare.
}
\keyword{datasets}
