% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_passionate_pilgrim}
\alias{the_passionate_pilgrim}
\title{Text of The Passionate Pilgrim}
\format{
A character vector of length 240. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_passionate_pilgrim
}
\description{
A character vector containing the complete text of:
The Passionate Pilgrim
part of which is attributed to William Shakespeare.
}
\keyword{datasets}
