% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_phoenix_and_the_turtle}
\alias{the_phoenix_and_the_turtle}
\title{Text of The Phoenix and the Turtle}
\format{
A character vector of length 94. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_phoenix_and_the_turtle
}
\description{
A character vector containing the complete text of:
The Phoenix and the Turtle
by William Shakespeare.
}
\keyword{datasets}
