% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_rape_of_lucrece}
\alias{the_rape_of_lucrece}
\title{Text of The Rape of Lucrece}
\format{
A character vector of length 2185. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_rape_of_lucrece
}
\description{
A character vector containing the complete text of:
The Rape of Lucrece
by William Shakespeare.
}
\keyword{datasets}
