% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tempest}
\alias{the_tempest}
\title{Text of The Tempest}
\format{
A character vector of length 3823. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_tempest
}
\description{
A character vector containing the complete text of:
The Tempest
by William Shakespeare.
}
\keyword{datasets}
