% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_two_noble_kinsmen}
\alias{the_two_noble_kinsmen}
\title{Text of The Two Noble Kinsmen}
\format{
A character vector of length 5188 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
the_two_noble_kinsmen
}
\description{
A character vector containing the complete text of:
The Two Noble Kinsmen
Partially attributed to William Shakespeare.
}
\keyword{datasets}
