% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timon_of_athens}
\alias{timon_of_athens}
\title{Text of Timon of Athens}
\format{
A character vector of length 2716 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
timon_of_athens
}
\description{
A character vector containing the complete text of:
Timon of Athens
by William Shakespeare.
}
\keyword{datasets}
