% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titus_andronicus}
\alias{titus_andronicus}
\title{Text of Titus Andronicus}
\format{
A character vector of length 2876. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
titus_andronicus
}
\description{
A character vector containing the complete text of:
Titus Andronicus
by William Shakespeare.
}
\keyword{datasets}
