% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troilus_and_cressida}
\alias{troilus_and_cressida}
\title{Text of Troilus and Cressida}
\format{
A character vector of length 6163. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
troilus_and_cressida
}
\description{
A character vector containing the complete text of:
Troilus and Cressida
by William Shakespeare.
}
\keyword{datasets}
