% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twelfth_night}
\alias{twelfth_night}
\title{Text of Twelfth Night}
\format{
A character vector of length 4483. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
twelfth_night
}
\description{
A character vector containing the complete text of:
Twelfth Night
by William Shakespeare.
}
\keyword{datasets}
