% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_gentlemen_of_verona}
\alias{two_gentlemen_of_verona}
\title{Text of Two Gentlemen of Verona}
\format{
A character vector of length 2412. Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
two_gentlemen_of_verona
}
\description{
A character vector containing the complete text of:
Two Gentlemen of Verona
by William Shakespeare.
}
\keyword{datasets}
