% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus_and_adonis}
\alias{venus_and_adonis}
\title{Text of Venus and Adonis}
\format{
A character vector of length 1431 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
venus_and_adonis
}
\description{
A character vector containing the complete text of:
Venus and Adonis
by William Shakespeare.
}
\keyword{datasets}
