% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winters_tale}
\alias{winters_tale}
\title{Text of A Winter's Tale}
\format{
A character vector of length 5015 Each line is <= 70 characters.
}
\source{
\url{http://www.gutenberg.org/files/100/100-0.txt}
}
\usage{
winters_tale
}
\description{
A character vector containing the complete text of:
A Winter's Tale
by William Shakespeare.
}
\details{
Exit, pursued by bear
}
\keyword{datasets}
